+++ src/decorators/services/injectable.decorator.ts
@@ -1,6 +1,9 @@
 import { CtorT } from '../../core/ctor';
+import { InjectableMetadata, InjectableMetadataSym } from '../../metadata/services/injectable';
 
-export function Injectable() {
-    return function(service: CtorT<any>) {
+export function Injectable<T>(meta?: InjectableMetadata<T>) {
+    meta = meta || {};
+    return function(injectable: CtorT<T>) {
+        Reflect.defineMetadata(InjectableMetadataSym, meta, injectable.prototype);
     }
 }