+++ src/metadata/services/index.ts
@@ -1,4 +1,5 @@
 
 
+export * from './injectable';
 export * from './metadata';
 export * from './service';