+++ src/metadata/services/injectable.ts
@@ -1,0 +1,6 @@
+import { CallbackSource } from '../server/provide';
+
+export type InjectableMetadata<T> = {
+    provide?: CallbackSource<T>
+}
+export const InjectableMetadataSym = Symbol.for('InjectableMetadata');