+++ src/metadata/services/service.ts
@@ -3,4 +3,4 @@
 export type ServiceMetadata = {
     
 }
-export const ServiceMetadataSym = Symbol.for('SeviceMetadata');
+export const ServiceMetadataSym = Symbol.for('ServiceMetadata');